Refactor dependency injections modules.
