+++ app/src/main/java/com/appunite/loudius/di/RepositoryModule.kt
@@ -1,11 +1,13 @@
 package com.appunite.loudius.di
 
+import com.appunite.loudius.domain.AuthRepository
+import com.appunite.loudius.domain.AuthRepositoryImpl
 import com.appunite.loudius.domain.PullRequestRepository
 import com.appunite.loudius.domain.PullRequestRepositoryImpl
+import com.appunite.loudius.domain.UserLocalDataSource
+import com.appunite.loudius.network.datasource.AuthDataSource
 import com.appunite.loudius.network.datasource.PullRequestDataSource
-import com.appunite.loudius.network.datasource.PullRequestsNetworkDataSource
 import com.appunite.loudius.network.datasource.UserDataSource
-import com.appunite.loudius.network.services.PullRequestsService
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -14,12 +16,7 @@ import javax.inject.Singleton
 
 @InstallIn(SingletonComponent::class)
 @Module
-object PullRequestModule {
-
-    @Provides
-    @Singleton
-    fun providePullRequestNetworkDataSource(service: PullRequestsService): PullRequestDataSource =
-        PullRequestsNetworkDataSource(service)
+object RepositoryModule {
 
     @Provides
     @Singleton
@@ -27,4 +24,11 @@ object PullRequestModule {
         dataSource: PullRequestDataSource,
         userDataSource: UserDataSource,
     ): PullRequestRepository = PullRequestRepositoryImpl(dataSource, userDataSource)
+
+    @Singleton
+    @Provides
+    fun provideAuthRepository(
+        authDataSource: AuthDataSource,
+        userLocalDataSource: UserLocalDataSource,
+    ): AuthRepository = AuthRepositoryImpl(authDataSource, userLocalDataSource)
 }
