+++ app/src/main/java/com/appunite/loudius/di/ServiceModule.kt
@@ -1,0 +1,31 @@
+package com.appunite.loudius.di
+
+import com.appunite.loudius.network.services.AuthService
+import com.appunite.loudius.network.services.PullRequestsService
+import com.appunite.loudius.network.services.UserService
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+import retrofit2.Retrofit
+
+@InstallIn(SingletonComponent::class)
+@Module
+object ServiceModule {
+
+    @Singleton
+    @Provides
+    fun provideAuthService(@AuthAPI retrofit: Retrofit): AuthService =
+        retrofit.create(AuthService::class.java)
+
+    @Singleton
+    @Provides
+    fun provideUserService(@BaseAPI retrofit: Retrofit): UserService =
+        retrofit.create(UserService::class.java)
+
+    @Singleton
+    @Provides
+    fun provideReposService(@BaseAPI retrofit: Retrofit): PullRequestsService =
+        retrofit.create(PullRequestsService::class.java)
+}
